move context tests to context
